1.自动生成dao：增加status,create_time,update_time列 2.修复自动生成的dao的bug。

+++ marketing-tool-service-impl/src/main/java/cn/rongcapital/mkt/service/impl/AudienceListDeleteServiceImpl.java
@@ -37,7 +37,7 @@ public class AudienceListDeleteServiceImpl implements AudienceListDeleteService{
 		AudienceListPartyMap audienceListPartyMap = new AudienceListPartyMap();
 		
 		audienceListPartyMap.setAudienceListId(audienceListId);
-		audienceListPartyMap.setStatus(ApiConstant.INT_ONE);
+		audienceListPartyMap.setStatus((byte)ApiConstant.INT_ONE);
 		audienceListPartyMap.setUpdateTime(new Date());
 		
 		BaseOutput result = new BaseOutput(ApiErrorCode.SUCCESS.getCode(),ApiErrorCode.SUCCESS.getMsg(),ApiConstant.INT_ZERO,null);
